requirements.txt
pyyaml
torch
torchvision
tensorboardX
mlflow
tensorboard